requirements.txt
click==8.0.2
cycler==0.10.0
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
